Adding CommonConditions & reconfigure tests.
