+++ HomeWork_17_SeleniumWD_Basic/src/test/java/HomeWorks/SeleniumPositiveTest.java
@@ -9,17 +9,8 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class SeleniumPositiveTest {
+public class SeleniumPositiveTest extends CommonConditions {
 
-    protected WebDriver driver;
-    private static final By MESSAGE = By.xpath("//div[@id='flash']");
-    private static final String ALERT_BASE_LOCATOR = "//button[@onclick='%s']";
-    private final String LOGIN_PAGE_URL = "https://the-internet.herokuapp.com/login";
-    @BeforeMethod
-    public void initBrowser() {
-        driver = new ChromeDriver();
-        driver.manage().window().maximize();
-    }
     @Test
     public void TestWithCorrectData() {
         driver.get("https://the-internet.herokuapp.com/login");
@@ -36,12 +27,5 @@ public class SeleniumPositiveTest {
         JavascriptExecutor js = (JavascriptExecutor) driver;
         String str = js.executeScript("return arguments[0].firstChild.textContent", el).toString();
         System.out.println("The message is: " + str);
-    }
-
-    @AfterMethod
-    public void closeDriver (){
-       if(driver != null) {
-           driver.quit();
-       }
     }
 }
